requirements.txt
requests
pyyaml
pandas
networkx